requirements.txt
pydantic==1.10.2
mypy==0.982
wemake-python-styleguide==0.17.0
python-telegram-bot==13.14
httpx==0.23.0
